requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.13.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
tvdb_v4_official==1.1.0
httpx==0.25.2
structlog==23.2.0
tenacity==8.2.3
prometheus-client==0.19.0
slowapi==0.1.9
python-dotenv==1.0.0